% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rocket_channels.R
\name{get_rocket_channels}
\alias{get_rocket_channels}
\title{Retrieve a list of all public/open channels in a Rocket.chat server}
\usage{
get_rocket_channels(url, user_id, auth_token)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}
}
\value{
The return of this function is a dataframe.
}
\description{
\code{get_rocket_channels} Gets a dataframe with all the channels in a Rocket.chat server.
}
\details{
This function connects to the Rocket.chat server and returns a dataframe
containing the name of the channel, the channels's ID, number of users and number of
messages exchanged there.

The function may be similar (some might say is equal) to `get_rocket_groups`,
but Rocket.chat has a few differences in the API regarding public/open channels
and private groups, this is the reason there are two functions.
}
\examples{

\dontrun{
get_rocket_channels(url        = 'http://localhost:8000/',
                    user_id    = '<ROCKET_USER_ID>',
                    auth_token = '<ROCKET_USER_AUTH_TOKEN>')
}
}
