% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rocket_credentials.R
\name{get_rocket_credentials}
\alias{get_rocket_credentials}
\title{Get Rocket.chat user credentials}
\usage{
get_rocket_credentials(url, user,
  password = rstudioapi::askForPassword())
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https)}

\item{user}{The username of login which will be used to extract ID and Token}

\item{password}{This argument can be set manually, but as default it opens a popup in RStudio for the user to type the password.}
}
\value{
Given a valid url, user and password, this function returns a named list
containing the user_id argument and the auth_token argument.
}
\description{
\code{get_rocket_credentials} Connects with the Rocket.chat REST API and returns the user ID and token.
}
\details{
This function connects with the Rocket.chat REST API to gather the users ID and token.
This sort of information will be needed to any function, since the "contract" between R
and the API of rocket depends on Token Authentication
}
\examples{

\dontrun{
get_rocket_credentials('http://localhost:8000/', user = 'me')
}
}
