% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rocket_groups_history.R
\name{get_rocket_groups_history}
\alias{get_rocket_groups_history}
\title{Retrieve a list of `n`` messages exchanged in private groups of a Rocket.chat server}
\usage{
get_rocket_groups_history(url, user_id, auth_token, group,
  n_messages = 0)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}

\item{group}{The name of the group to get the history from}

\item{n_messages}{The number of messages to be returned (default is 0, which brings the complete history)}
}
\value{
The return of this function is a dataframe.
}
\description{
\code{get_rocket_groups_history} Gets a dataframe with the messages
exchanged in a specific group of a Rocket.chat server.
}
\details{
This function returns a dataframe containing `n` messages exchanged in
a private group in a Rocket.chat server.

The return of this function contains the content of the message, who sent it (username and name),
the ID of the sender, and a time stamp.

The function may be similar (some might say is equal) to `get_rocket_channels_history`,
but Rocket.chat has a few differences in the API regarding public/open channels
and private groups, this is the reason there are two functions.
}
\examples{

\dontrun{
get_rocket_groups_history(url        = 'http://localhost:8000/',
                          user_id    = '<ROCKET_USER_ID>',
                          auth_token = '<ROCKET_USER_AUTH_TOKEN>',
                          group = 'bot_tests',
                          n_messages = 0)
}
}
