% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rocket_room_exists.R
\name{get_rocket_room_exists}
\alias{get_rocket_room_exists}
\title{Validates whether a room exists or not}
\usage{
get_rocket_room_exists(url, user_id, auth_token, room_id)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}

\item{room_id}{The id of the room to be checked}
}
\value{
The return of this function is a boolean
}
\description{
\code{get_rocket_room_exists} Validates whether a room exists
}
\details{
This function returns a logical response to whether a room exists or not
}
\examples{

\dontrun{
get_rocket_room_exists(url        = 'http://localhost:8000/',
                       user_id    = '<ROCKET_USER_ID>',
                       auth_token = '<ROCKET_USER_AUTH_TOKEN>',
                       room_id    = '<ROOM_ID>')
}
}
