% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rocket_users.R
\name{get_rocket_users}
\alias{get_rocket_users}
\title{Retrieve a list of all users in a Rocket.chat server}
\usage{
get_rocket_users(url, user_id, auth_token)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}
}
\value{
The return of this function is a dataframe.
}
\description{
\code{get_rocket_users} Gets a dataframe with all the users in a Rocket.chat server.
}
\details{
This is a simple function, that connects with the Rocket.chat server
and returns a dataframe with information about all users registered.

The return comprises a few information about the users such as:
* username  = Username used to login.
* name      = Actual name of the user.
* user_id   = ID used to identify other users within Rocket.chat.
* is_ldap   = If this registry comes from LDAP servers or not.
* is_active = If this user is still active.
* type      = Can be one of two, or user, or bot.
}
\examples{

\dontrun{
get_rocket_users(url        = 'http://localhost:8000/',
                 user_id    = '<ROCKET_USER_ID>',
                 auth_token = '<ROCKET_USER_AUTH_TOKEN>')
}
}
