% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_rocket_attachment.R
\name{send_rocket_attachment}
\alias{send_rocket_attachment}
\title{Sends an attachment to a rocket user/channel}
\usage{
send_rocket_attachment(url, user_id, auth_token, receiver, filename)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}

\item{receiver}{The id of the user/channel/group that will receive the attachment (with `@` for users and `#` for channels and users)}

\item{filename}{The path of the file to be uploaded.}
}
\value{
This function only returns the answer of the POST request
}
\description{
\code{send_rocket_attachment} Sends an attachment to a user or channel or group
}
\details{
This function only returns the status code of the POST request to upload an
attachment for any user or channel or group in a Rocket.chat server.
}
\examples{

\dontrun{
send_rocket_attachment(url        = 'http://localhost:8000/',
                       user_id    = '<ROCKET_USER_ID>',
                       auth_token = '<ROCKET_USER_AUTH_TOKEN>',
                       receiver   = '#general',
                       filename   = 'path_to_file')
}
}
