% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_rocket_message.R
\name{send_rocket_message}
\alias{send_rocket_message}
\title{Send Rocket.chat messages to users, channels or groups}
\usage{
send_rocket_message(url, user_id, auth_token, receiver, message)
}
\arguments{
\item{url}{The url of the rocket.chat server (has to be complete, with http or https).}

\item{user_id}{The users ID retrieved from `get_rocket_credentials`.}

\item{auth_token}{The users authentication Token retrieved from `get_rocket_credentials`.}

\item{receiver}{The receiver of the message (if user mentioned with `@`, otherwise use `#`)}

\item{message}{A text string containing the text to be sent}
}
\value{
This function has no return since it posts the message to the server
}
\description{
\code{send_rocket_message} Uses the users ID and Token to communicate with Rocket.chat server and send messages to specific groups, users or channels
}
\details{
This function connects with the Rocket.chat with users ID and Token and send
a text message to a receiver.

To send a message to specific user, is essential to use the `@` before the user name.
To send a message to groups or channels, you should use `#`
}
\examples{

\dontrun{
send_rocket_message(url        = 'http://localhost:8000/',
                    user_id    = '<ROCKET_USER_ID>',
                    auth_token = '<ROCKET_USER_AUTH_TOKEN>',
                    receiver   = '#general',
                    message    = 'hello')
}
}
